rename MainPage to WelcomePage
